backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.135.3"}
psycopg = {extras = ["binary"], version = "^3.3.3"}
sqlalchemy = "^2.0.49"
aiosqlite = "^0.22.1"
httpx = "^0.28.1"
python-multipart = "^0.0.22"
pydantic-settings = "^2.13.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.15.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
